openthairag/requirements.txt
flask
flask-cors
pymilvus==2.4.7
transformers
torch
scikit-learn
gunicorn
numpy
pymongo
openai
requests
dnspython==2.3.0
flask_swagger_ui
flask-jwt-extended
pyjwt<2.10
python-socketio[client]
timezonefinder==6.5.8
geopy
accelerate
marshmallow==3.26.1
watchgod 
mammoth
line-bot-sdk==3.17.1

pandas
huggingface-hub
sentence-transformers
Pillow
llama-index
llama-index-embeddings-huggingface 
llama-index-vector-stores-qdrant 
qdrant-client 
python-dotenv 
PyPDF2

crewai 
gspread 
oauth2client
mysql-connector-python
together